# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserPublic {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserPublic;
}

export interface CreateContractBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 20 */
  content: string;
}

export interface ContractSummary {
  id: number;
  title: string;
  fileName?: string | null;
  createdAt: string;
  analysisStatus?: string | null;
}

export interface Contract {
  id: number;
  title: string;
  content: string;
  fileName?: string | null;
  fileSize?: number | null;
  createdAt: string;
}

export interface RightItem {
  title: string;
  description: string;
  citations: string[];
}

export type AlertItemSeverity =
  (typeof AlertItemSeverity)[keyof typeof AlertItemSeverity];

export const AlertItemSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type AlertItemAffectedParty =
  (typeof AlertItemAffectedParty)[keyof typeof AlertItemAffectedParty];

export const AlertItemAffectedParty = {
  employee: "employee",
  employer: "employer",
  both: "both",
} as const;

export interface AlertItem {
  severity: AlertItemSeverity;
  clause: string;
  issue: string;
  affectedParty: AlertItemAffectedParty;
  recommendation: string;
  citations: string[];
}

export type CitationItemSource =
  (typeof CitationItemSource)[keyof typeof CitationItemSource];

export const CitationItemSource = {
  law: "law",
  regulation: "regulation",
} as const;

export interface CitationItem {
  ref: string;
  source: CitationItemSource;
  articleNumber: string;
  title: string;
  excerpt: string;
}

export interface ContractAnalysis {
  id: number;
  contractId: number;
  status: string;
  summary?: string | null;
  contractType?: string | null;
  partiesDescription?: string | null;
  employeeRights?: RightItem[] | null;
  employerRights?: RightItem[] | null;
  alerts?: AlertItem[] | null;
  citations?: CitationItem[] | null;
  errorMessage?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface ContractWithAnalysis {
  contract: Contract;
  analysis?: ContractAnalysis | null;
}

export interface LawArticle {
  id: number;
  source: string;
  bookTitle?: string | null;
  chapterTitle?: string | null;
  sectionTitle?: string | null;
  articleNumber: string;
  articleLabel?: string | null;
  content: string;
  summary?: string | null;
}

export type UploadContractBody = {
  title?: string;
  file: Blob;
};

export type SearchLawArticlesParams = {
  q?: string;
  source?: SearchLawArticlesSource;
  limit?: number;
};

export type SearchLawArticlesSource =
  (typeof SearchLawArticlesSource)[keyof typeof SearchLawArticlesSource];

export const SearchLawArticlesSource = {
  law: "law",
  regulation: "regulation",
} as const;