# lib/api-zod/src/generated/types/alertItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertItemAffectedParty } from "./alertItemAffectedParty";
import type { AlertItemSeverity } from "./alertItemSeverity";

export interface AlertItem {
  severity: AlertItemSeverity;
  clause: string;
  issue: string;
  affectedParty: AlertItemAffectedParty;
  recommendation: string;
  citations: string[];
}