# lib/api-zod/src/generated/types/citationItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { CitationItemSource } from "./citationItemSource";

export interface CitationItem {
  ref: string;
  source: CitationItemSource;
  articleNumber: string;
  title: string;
  excerpt: string;
}