# lib/api-zod/src/generated/types/contractWithAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { Contract } from "./contract";
import type { ContractAnalysis } from "./contractAnalysis";

export interface ContractWithAnalysis {
  contract: Contract;
  analysis?: ContractAnalysis | null;
}