# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 20 */
  content: string;
}