# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Molem - Saudi Labor Law Contract Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  name: string;
}